refactor SelectionEditor into own file
